requirements.txt
-i https://pypi.org/simple
click==7.0
dominate==2.4.0
flask-bootstrap4==4.0.2
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.11.1
markupsafe==1.1.1
visitor==0.1.3
werkzeug==1.0.0
